models: use BigInteger for id/FK columns to match DB BIGINT (fixes audit_log FK)
The legacy Grails/GORM schema declares every id primary key and *_id
foreign key as BIGINT. The ORM previously declared them as Integer, which
works for CRUD but caused the audit_log FK (user_id INT -> user.id BIGINT)
to fail with errno 150 when running migration 001.

Sweep all models to BigInteger for id PKs and ForeignKey columns, verified
against docs/baco_import_all.sql. Business/int columns (hours, counts,
thresholds) stay Integer, as do the genuine int columns the dump shows:
court_order.order_id, rental_log.transaction_id, and audit_log.id/club_id
(the new table's own INT columns). Migration 001 audit_log.user_id also
changed INT -> BIGINT.

+++ backend/app/models/audit.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from sqlalchemy import Column, DateTime, ForeignKey, Integer, String, Text
+from sqlalchemy import BigInteger, Column, DateTime, ForeignKey, Integer, String, Text
 
 from app.database import Base
 
@@ -13,11 +13,11 @@ class AuditLog(Base):
     """
     __tablename__ = "audit_log"
 
-    id = Column(Integer, primary_key=True)
+    id = Column(Integer, primary_key=True)        # audit_log.id is INT in migration 001 (its own new table)
     created_at = Column(DateTime, nullable=False, default=datetime.now)  # server local time
-    user_id = Column(Integer, ForeignKey("user.id"), nullable=True)
+    user_id = Column(BigInteger, ForeignKey("user.id"), nullable=True)   # BigInteger to match user.id (BIGINT)
     user_name = Column(String(255), nullable=False, default="")
-    club_id = Column(Integer, nullable=True)      # null = global action (e.g. full rebuild)
+    club_id = Column(Integer, nullable=True)      # null = global action (e.g. full rebuild); no FK, denormalized
     club_name = Column(String(255), nullable=True)
     action = Column(String(64), nullable=False)   # stable code, e.g. "schedule.save"
     summary = Column(String(512), nullable=False, default="")  # human Hebrew one-liner
